algorithm/GearDetection/requirements.txt
# 齿轮缺陷检测平台依赖包

# Web框架
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.1

# 深度学习
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
Pillow>=8.2.0

# 工具
python-dotenv==0.19.0